+++ src/main/java/edu/teco/smartlambda/monitoring/MonitoringService.java
@@ -47,7 +47,7 @@ public class MonitoringService {
 	 * @param CPUTime              the lambda used
 	 * @param executionReturnValue returnValue or exception of lambda
 	 */
-	public void onLambdaExecutionEnd(final AbstractLambda lambda, final int CPUTime, final ExecutionReturnValue executionReturnValue) {
+	public void onLambdaExecutionEnd(final AbstractLambda lambda, final long CPUTime, final ExecutionReturnValue executionReturnValue) {
 		this.monitoringEvent.setCPUTime(CPUTime);
 		this.monitoringEvent.setDuration(Calendar.getInstance().getTimeInMillis() - this.monitoringEvent.getTime().getTimeInMillis());
 		if (executionReturnValue.isException()) {
